feat: Server class for handles
